Do not set media as disabled when media is not available
The audio and video are explicitly set as disabled before they are set
as not available, so there is no need to set them as disabled again.
Moreover, even if they were not explicitly set as disabled it would not
be really necessary to disable the audio and video when they are not
available.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>